+++ lib/expressions.dart
@@ -1,4 +1,4 @@
-library;
+library expressions;
 
 export 'src/expressions/evaluator.dart';
 export 'src/expressions/expressions.dart';