+++ test/expressions/async_evaluator_test.dart
@@ -121,23 +121,19 @@ void main() {
         stream.listen((v) => current = v, onDone: () => done = true);
 
         controllerX.add(40);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, null);
 
         controllerY.add(60);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, false);
 
         controllerY.add(20);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, true);
 
         controllerX.add(10);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, false);
 
         await async.flushMicrotasksUntil(controllerX.close());
@@ -280,39 +276,30 @@ void main() {
 
         dynamic current;
         var done = false;
-        stream.listen(
-          (v) => current = v,
-          onDone: () => done = true,
-        );
+        stream.listen((v) => current = v, onDone: () => done = true);
 
         controllerX.add('y');
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, null);
 
         controllerY.add(1);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, 1);
 
         controllerX.add('z');
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, 1);
 
         controllerZ.add(2);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, 2);
 
         controllerY.add(3);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, 2);
 
         controllerZ.add(4);
-        // async.flushMicrotasks();
-        await async.unblock();
+        async.flushMicrotasks();
         expect(current, 4);
 
         await async.flushMicrotasksUntil(controllerX.close());
@@ -413,9 +400,8 @@ void main() {
 }
 
 FutureOr<T> fakeAsync<T>(
-  FutureOr<T> Function(fake_async.FakeAsync async) callback, {
-  DateTime? initialTime,
-}) {
+    FutureOr<T> Function(fake_async.FakeAsync async) callback,
+    {DateTime? initialTime}) {
   final async = fake_async.FakeAsync(initialTime: initialTime);
   final f = async.run(callback);
   if (f is Future<T>) {
@@ -428,8 +414,7 @@ extension FakeAsyncX on fake_async.FakeAsync {
   Future<T> flushMicrotasksUntil<T>(Future<T> f) async {
     var isDone = false;
     f = f.whenComplete(
-      () => isDone = true,
-    ); // check if all work in body has been done
+        () => isDone = true); // check if all work in body has been done
     while (!isDone) {
       // flush the microtasks in real async zone
       await unblock();