+++ test/templates/built_in_objects_test.dart
@@ -178,6 +178,14 @@ void main() {
         template.process(context: context),
         '2022-02-07',
       );
+
+      template = Template(
+        value: r'${DateTime([2022, 02, 07]).format("yyyy-MM-dd")}',
+      );
+      expect(
+        template.process(context: context),
+        '2022-02-07',
+      );
     });
 
     test('now', () {
@@ -349,6 +357,60 @@ void main() {
         'John',
       );
     });
+
+    test('list', () {
+      expect(
+        Template(value: r'${x.path("$[1]")}').process(
+          context: {
+            'x': [
+              'foo',
+              'bar',
+            ],
+          },
+        ),
+        'bar',
+      );
+    });
+    test('map', () {
+      expect(
+        Template(value: r'${x.path("$.bar")}').process(context: {
+          'x': {
+            'foo': 'foo',
+            'bar': 'bar',
+          },
+        }),
+        'bar',
+      );
+    });
+    test('json string', () {
+      expect(
+        Template(value: r'${x.path("$.bar")}').process(
+          context: {
+            'x': '''
+            {
+              'foo': 'foo',
+              'bar': 'bar',
+            }
+          '''
+          },
+        ),
+        'bar',
+      );
+    });
+
+    test('yaml string', () {
+      expect(
+        Template(value: r'${x.path("$.bar")}').process(
+          context: {
+            'x': '''
+foo: foo
+bar: bar
+'''
+          },
+        ),
+        'bar',
+      );
+    });
   });
 
   group('List<int>', () {
@@ -378,6 +440,56 @@ void main() {
           context: {'input': utf8.encode(input)},
         ),
         input,
+      );
+    });
+  });
+
+  group('NumberFormat', () {
+    test('format', () {
+      expect(
+        Template(value: r'${NumberFormat("#,###.00").format(x)}')
+            .process(context: {
+          'x': 1234,
+        }),
+        '1,234.00',
+      );
+      expect(
+        Template(value: r'${NumberFormat("#,###.00").format(x)}')
+            .process(context: {
+          'x': 1234.0,
+        }),
+        '1,234.00',
+      );
+      expect(
+        Template(value: r'${x.format("#,###.00")}').process(context: {
+          'x': 1234.0,
+        }),
+        '1,234.00',
+      );
+      expect(
+        Template(value: r'${x.format("#,###.00")}').process(context: {
+          'x': 1234,
+        }),
+        '1,234.00',
+      );
+    });
+
+    test('parse', () {
+      expect(
+        Template(value: r'${x.toInt()}').evaluate(context: {'x': '1234.45'}),
+        1234,
+      );
+      expect(
+        Template(value: r'${x.toDouble()}').evaluate(context: {'x': '1234.45'}),
+        1234.45,
+      );
+      expect(
+        Template(value: r'${x.toInt()}').evaluate(context: {'x': 'foo'}),
+        null,
+      );
+      expect(
+        Template(value: r'${x.toDouble()}').evaluate(context: {'x': 'foo'}),
+        null,
       );
     });
   });